requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
openai>=0.28.0
solana>=0.36.6
httpx>=0.28.1
solders>=0.26.0
spl-token==4.0.0
requests==2.31.0
numpy
pandas
scikit-learn
tensorflow